Fix coloring
