+++ backend/config/config.go
@@ -108,7 +108,7 @@ func LoadConfig() *Config {
 			dbURL = built
 		} else {
 			// Safe local default for dev
-			dbURL = "postgres://postgres:postgres@localhost:5432/leaflock?sslmode=prefer"
+			dbURL = "postgres://postgres:postgres@localhost:5432/leaflock?sslmode=prefer" // secretlint-disable-line
 		}
 	}
